feat: created components improving organization
