+++ src/components/Project.jsx
@@ -1,0 +1,127 @@
+import site from '../assets/site.png'
+import site2 from '../assets/site2.png'
+import site3 from '../assets/site3.png'
+
+const Project = () => {
+  return (
+              <section id="projetos" className="mt-16 ">
+                <h2 className="text-3xl font-semibold text-green-500 mb-6 text-center">Projetos</h2>
+                <div className="grid grid-cols-1 md:grid-cols-3 gap-6">
+    
+    
+                  <div className=" bg-white  rounded-lg shadow-sm dark:bg-gray-800  backdrop-blur-xl  ring-1 ring-white/20 transition-all duration-300 hover:backdrop-blur-2xl hover:shadow-[0_8px_32px_0_rgba(31,38,135,0.37)] hover:scale-105 hover:border-white/50" >
+                    <a href="#">
+                      <img className="rounded-t-lg" src={site} alt="Foto de Jocely Marques" />
+                    </a>
+                    <div className="flex flex-col p-5 gap-5">
+                      <a href="#">
+                        <h5 className="mb-2 text-2xl font-bold tracking-tight text-gray-900 dark:text-white">Projeto 1</h5>
+                      </a>
+                      <div className="flex flex-wrap gap-2 mb-5">
+                        <span className="inline-flex w-fit items-center rounded-full bg-violet-400/20 px-2 py-1 text-xs font-medium text-gray-400 inset-ring inset-ring-gray-400/20">
+                          Reactjs
+                        </span>
+                        <span className="inline-flex w-fit items-center rounded-full bg-orange-400/20 px-2 py-1 text-xs font-medium text-gray-400 inset-ring inset-ring-gray-400/20">
+                          Laravel
+                        </span>
+                        <span className="inline-flex w-fit items-center rounded-full bg-gray-400/20 px-2 py-1 text-xs font-medium text-gray-400 inset-ring inset-ring-gray-400/20">
+                          SQL
+                        </span>
+                        <span className="inline-flex w-fit items-center rounded-full bg-yellow-400/20 px-2 py-1 text-xs font-medium text-gray-400 inset-ring inset-ring-gray-400/20">
+                          HTML
+                        </span>
+                        <span className="inline-flex w-fit items-center rounded-full bg-blue-400/20 px-2 py-1 text-xs font-medium text-gray-400 inset-ring inset-ring-gray-400/20">
+                          CSS
+                        </span>
+                      </div>
+                      <div className="flex justify-evenly gap-3">
+                        <a href="#" className="inline-flex items-center px-10 py-2 text-sm font-medium text-center text-white bg-blue-700 rounded-lg hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800">
+                          Preview
+                        </a>
+                        <a href="#" className="inline-flex items-center px-10 py-2 text-sm font-medium text-center text-white bg-gray-700 rounded-lg hover:bg-gray-800 focus:ring-4 focus:outline-none focus:ring-gray-300 dark:bg-gray-600 dark:hover:bg-gray-700 dark:focus:ring-gray-800">
+                          GitHub
+                        </a>
+                      </div>
+                    </div>
+                  </div>
+    
+                  <div className=" bg-white  rounded-lg shadow-sm dark:bg-gray-800  backdrop-blur-xl  ring-1 ring-white/20 transition-all duration-300 hover:backdrop-blur-2xl hover:shadow-[0_8px_32px_0_rgba(31,38,135,0.37)] hover:scale-105 hover:border-white/50" >
+                    <a href="#">
+                      <img className="rounded-t-lg" src={site2} alt="Foto de Jocely Marques" />
+                    </a>
+                    <div className="flex flex-col p-5 gap-5">
+                      <a href="#">
+                        <h5 className="mb-2 text-2xl font-bold tracking-tight text-gray-900 dark:text-white">Projeto 2</h5>
+                      </a>
+                      <div className="flex flex-wrap gap-2 mb-5">
+                        <span className="inline-flex w-fit items-center rounded-full bg-violet-400/20 px-2 py-1 text-xs font-medium text-gray-400 inset-ring inset-ring-gray-400/20">
+                          Reactjs
+                        </span>
+                        <span className="inline-flex w-fit items-center rounded-full bg-orange-400/20 px-2 py-1 text-xs font-medium text-gray-400 inset-ring inset-ring-gray-400/20">
+                          Laravel
+                        </span>
+                        <span className="inline-flex w-fit items-center rounded-full bg-gray-400/20 px-2 py-1 text-xs font-medium text-gray-400 inset-ring inset-ring-gray-400/20">
+                          SQL
+                        </span>
+                        <span className="inline-flex w-fit items-center rounded-full bg-yellow-400/20 px-2 py-1 text-xs font-medium text-gray-400 inset-ring inset-ring-gray-400/20">
+                          HTML
+                        </span>
+                        <span className="inline-flex w-fit items-center rounded-full bg-blue-400/20 px-2 py-1 text-xs font-medium text-gray-400 inset-ring inset-ring-gray-400/20">
+                          CSS
+                        </span>
+                      </div>
+                      <div className="flex justify-evenly gap-3">
+                        <a href="#" className="inline-flex items-center px-10 py-2 text-sm font-medium text-center text-white bg-blue-700 rounded-lg hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800">
+                          Preview
+                        </a>
+                        <a href="#" className="inline-flex items-center px-10 py-2 text-sm font-medium text-center text-white bg-gray-700 rounded-lg hover:bg-gray-800 focus:ring-4 focus:outline-none focus:ring-gray-300 dark:bg-gray-600 dark:hover:bg-gray-700 dark:focus:ring-gray-800">
+                          GitHub
+                        </a>
+                      </div>
+                    </div>
+                  </div>
+    
+    
+                  <div className=" bg-white  rounded-lg shadow-sm dark:bg-gray-800  backdrop-blur-xl  ring-1 ring-white/20 transition-all duration-300 hover:backdrop-blur-2xl hover:shadow-[0_8px_32px_0_rgba(31,38,135,0.37)] hover:scale-105 hover:border-white/50" >
+                    <a href="#">
+                      <img className="rounded-t-lg" src={site3} alt="Foto de Jocely Marques" />
+                    </a>
+                    <div className="flex flex-col p-5 gap-5">
+                      <a href="#">
+                        <h5 className="mb-2 text-2xl font-bold tracking-tight text-gray-900 dark:text-white">Projeto 3</h5>
+                      </a>
+                      <div className="flex flex-wrap gap-2 mb-5">
+                        <span className="inline-flex w-fit items-center rounded-full bg-violet-400/20 px-2 py-1 text-xs font-medium text-gray-400 inset-ring inset-ring-gray-400/20">
+                          Reactjs
+                        </span>
+                        <span className="inline-flex w-fit items-center rounded-full bg-orange-400/20 px-2 py-1 text-xs font-medium text-gray-400 inset-ring inset-ring-gray-400/20">
+                          Laravel
+                        </span>
+                        <span className="inline-flex w-fit items-center rounded-full bg-gray-400/20 px-2 py-1 text-xs font-medium text-gray-400 inset-ring inset-ring-gray-400/20">
+                          SQL
+                        </span>
+                        <span className="inline-flex w-fit items-center rounded-full bg-yellow-400/20 px-2 py-1 text-xs font-medium text-gray-400 inset-ring inset-ring-gray-400/20">
+                          HTML
+                        </span>
+                        <span className="inline-flex w-fit items-center rounded-full bg-blue-400/20 px-2 py-1 text-xs font-medium text-gray-400 inset-ring inset-ring-gray-400/20">
+                          CSS
+                        </span>
+                      </div>
+                      <div className="flex justify-evenly gap-3">
+                        <a href="#" className="inline-flex items-center px-10 py-2 text-sm font-medium text-center text-white bg-blue-700 rounded-lg hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800">
+                          Preview
+                        </a>
+                        <a href="#" className="inline-flex items-center px-10 py-2 text-sm font-medium text-center text-white bg-gray-700 rounded-lg hover:bg-gray-800 focus:ring-4 focus:outline-none focus:ring-gray-300 dark:bg-gray-600 dark:hover:bg-gray-700 dark:focus:ring-gray-800">
+                          GitHub
+                        </a>
+                      </div>
+                    </div>
+                  </div>
+    
+    
+                </div>
+              </section>
+  )
+}
+
+export default Project;
